Change ROI in MainWindow  Changes to be committed: 	modified:   codes/AbstractData/Overlay.cxx 	modified:   codes/AbstractInteractorStyles/InteractorStyleROI.cxx 	modified:   codes/Module/ModuleWidget.cxx 	modified:   codes/Module/ModuleWidget.h 	modified:   codes/core/MainWindow.cxx 	modified:   codes/core/MainWindow.h 	modified:   codes/core/MyPlaneWidget.cxx

+++ codes/Module/ModuleWidget.h
@@ -56,7 +56,7 @@ public slots:
 	void slotSelectROI();
 	void slotResetROI();
 	void slotChangeOpacity(int opactiy);
-	void slotChangeROI(double* bound);
+	void slotChangeROI(int* bound);
 	//void slotAddLumen();
 	//void slotAddVesselWall();
 	void GenerateReportPushButtonVisible();
